Add support for reading the file contents from stdin
Whenever that's set, the full output is always going to get printed to
stdout.

This is mainly intended for integration with editors, but it's made in a
way that is compatible with previous setups.

This includes a flag for the file name, I'm using `--stdin-display-name`
which matches what flake8 uses.

Closes #12.